Fix yarn lock file
Some changes in the `yarn.lock` file lead to a broken template page
which would display the error `Type error: t is not a function`

This only occurred when yarn was then started with a frozen lockfile and
not when running in developer mode.

It was discovered that replacing the yarn file with the contents from
the original weaveworks github and then replacing the weaveworks NPM
repos with the correct choclab one fixed the issue at least until I
can spend some time upgrading dependencies properly.