Templating functionality
In the original weave EE, templating does not apply to paths. This is
an issue if you wish to use a custom directory structure for your
installation.

This change introduces the capability of templating paths and defining
custom mappings to be injected into the templates.

Secondary to this, it is now possible to automatically populate
dropdowns in your templates with resources available inside the
management cluster. This is useful for presenting filtered lists of
namespaces or Giant Swarm Organizations as well as other types you wish
to use in your templates.