Upgrade UI components
In order to modernise the Weave EE UI it is necessary to upgrade the
libraries required by the interface.

This change introduces the following:

- Merge UI code from Weave GitOps OSS version - this is loaded into
  `ui/src/weave` as a first step
- Upgrade React from version 17 to version 18
- Upgrade React-Router-Dom from version 5 to version 6
- Upgrade Material UI from version 4 to version 5
- Upgrade all other dependencies to the latest equivelants

Due to a bug in this library it is currently included as a child library
under the `jslib` folder. I hope to remove this once all other components
are upgraded to meet the requirements of the upgrades attached to this
build.

Following this change, a number of patches are made to the code under
`ui/src` to ensure compatibility with the upgraded libraries. This
includes mainly:

- Replacing routing entries to match those used for
  `react-router-dom:v6`
- Upgrading styling to meet the new styles for `mui:v5`
- Fixing ESLint errors introduced by migrating to ES6
- Most, if not all style issues are updated against the newer version of
  material-ui using the existing `styled-components` backend
- There is a bug on the `profiles` section of creating from template
  whereby the form slows down dramatically due to the number of profiles
  loaded - I do not know if this was pre-existing

When migrating, initially there were a lot of issues running the unit
tests, many of which resulted in exceptionally poor performance and
memory leaks. This has resulted in re-structuring how the unit tests are
executed but only relative minor changes have been made to the tests
themselves to make them compatibile with the upgraded packages.

When building outside of the container, the build worked fine, however
when adding into a container this caused a runtime exception due to a
broken import for `yaml` module which resulted in a blank page at any
point in the application.

In order to rectify this, additional aliases have been included for
dependencies required by the `yaml` and `react-markdown` libraries.